# src/ovc/system_atlas/generated/atlas_openapi_v0_1.json
{"components":{"schemas":{"AtlasEnvelope":{"additionalProperties":false,"properties":{"completeness_profile":{"type":"string"},"data":{},"graph_generation":{"type":"string"},"query_policy_version":{"type":"string"},"repository_tree":{"type":"string"},"schema":{"const":"ovc-atlas-api-envelope/v1"},"security_visibility":{"items":{"type":"string"},"type":"array"},"warnings":{"items":{"type":"string"},"type":"array"},"write_effect":{"const":"NONE"}},"required":["schema","graph_generation","repository_tree","query_policy_version","completeness_profile","security_visibility","warnings","write_effect","data"],"type":"object"},"AtlasQueryRequest":{"additionalProperties":false,"properties":{"changed_entity_ids":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}]},"direction":{"anyOf":[{"type":"string"},{"type":"null"}]},"entity_id":{"anyOf":[{"type":"string"},{"type":"null"}]},"family":{"type":"string"},"max_depth":{"anyOf":[{"type":"integer"},{"type":"null"}]},"object_id":{"anyOf":[{"type":"string"},{"type":"null"}]},"predicates":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}]},"relationship_families":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}]},"start_id":{"anyOf":[{"type":"string"},{"type":"null"}]},"term":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["family"],"type":"object"},"AtlasViewRequest":{"additionalProperties":false,"properties":{"entity_ids":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}]},"maximum_entities":{"default":200,"maximum":500,"minimum":1,"type":"integer"}},"type":"object"}}},"info":{"description":"Generation-bound read-only Atlas API. POST routes are query semantics with write_effect=NONE.","title":"OVC System Atlas Read-Only API","version":"0.1.0"},"openapi":"3.1.0","paths":{"/api/v1/atlas/meta":{"get":{"operationId":"atlasMeta","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtlasEnvelope"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtlasEnvelope"}}},"description":"Query Rejected"}}}},"/api/v1/atlas/query":{"post":{"operationId":"atlasQuery","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtlasQueryRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtlasEnvelope"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtlasEnvelope"}}},"description":"Query Rejected"}}}},"/api/v1/atlas/view":{"post":{"operationId":"atlasView","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtlasViewRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtlasEnvelope"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AtlasEnvelope"}}},"description":"Query Rejected"}}}}}}